move util file to test util folder
