+++ src/integrationTest/codewhisperer/securityScan.test.ts
@@ -10,7 +10,7 @@ import * as CodeWhispererConstants from '../../codewhisperer/models/constants'
 import * as path from 'path'
 import * as testutil from '../../test/testUtil'
 import * as fs from 'fs-extra'
-import { setValidConnection } from './util'
+import { setValidConnection } from '../util/codewhispererUtil'
 import { resetCodeWhispererGlobalVariables } from '../../test/codewhisperer/testUtil'
 import { getTestWorkspaceFolder } from '../integrationTestsUtilities'
 import { closeAllEditors } from '../../test/testUtil'
